+++ catguild-system/src/main/java/cn/catguild/system/infrastructure/idgeneration/strategy/Snowflake.java
@@ -1,4 +1,4 @@
-package cn.catguild.system.infrastructure.id.strategy;
+package cn.catguild.system.infrastructure.idgeneration.strategy;
 
 import java.io.Serial;
 import java.io.Serializable;
